Ensure su is owned by root to avoid setgid issues
This occurs when using [user
namespaces](https://docs.docker.com/engine/security/userns-remap/) in
the Docker daemon, and running transmission in the host user namespace
(via `--userns host`).

The error manifests as:
```
setgid: Operation not permitted
```

This will happen even when using `--privileged` or `--add-cap`, as `su`
will have been updated as part of `apk upgrade`, and this will now mean
that `su` is owned by the user-namespaced `root` user, e.g.
231700:231700.

When running the container in the host's user namespace, `root` is now
the normal 0:0, which means `su` is now no longer eligible for `setgid`.